"Change: Change the position of some package"
- The adapter and listener packages have been moved into the presentation package.

close: #40
