+++ app/src/main/java/com/app/service/parking/di/RepositoryModules.kt
@@ -1,12 +1,11 @@
 package com.app.service.parking.di
 
 import com.app.service.parking.global.App.Companion.context
-import com.app.service.parking.model.repository.local.db.AppDB
-import com.app.service.parking.model.repository.local.repository.FavoriteRepository
+import com.app.service.parking.model.db.AppDB
+import com.app.service.parking.model.repository.local.FavoriteRepository
 import com.app.service.parking.model.repository.remote.ParkingLotRepository
 import com.app.service.parking.model.repository.remote.ReviewRepository
 import com.app.service.parking.model.repository.remote.UserRepository
-import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
 
 val userRepositoryModule = module {
@@ -19,7 +18,7 @@ val parkingRepositoryModule = module {
 
 val favoriteRepositoryModule = module {
     factory {
-        FavoriteRepository(AppDB.getDatabase(context!!))
+        FavoriteRepository.getInstance(AppDB.getDatabase(context!!))
     }
 }
