+++ app/src/main/java/com/app/service/parking/model/db/dao/DAOFavorite.kt
@@ -1,8 +1,8 @@
-package com.app.service.parking.model.repository.local.dao
+package com.app.service.parking.model.db.dao
 
 import androidx.lifecycle.LiveData
 import androidx.room.*
-import com.app.service.parking.model.repository.local.DBConst
+import com.app.service.parking.model.db.DBConst
 import com.app.service.parking.model.repository.entity.Favorite
