+++ app/src/main/java/com/app/service/parking/presentation/listener/POIItemClickListener.kt
@@ -1,4 +1,4 @@
-package com.app.service.parking.listener
+package com.app.service.parking.presentation.listener
 
 import net.daum.mf.map.api.MapPOIItem
