+++ app/src/main/java/com/app/service/parking/presentation/listener/RecyclerItemClickListener.kt
@@ -1,4 +1,4 @@
-package com.app.service.parking.listener
+package com.app.service.parking.presentation.listener
 
 interface RecyclerItemClickListener {
     // 어떤 버튼을 클릭했는지 구분해야하는 경우엔 resId를 지정한다.
